requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.0.267
openai>=1.3.0
jsonpatch>=1.33,<2.0
langchain-community
langchain_openai
# Template handling
jinja2>=3.1.2

# Document generation
python-docx>=0.8.11
python-dotenv==1.0.1

# UI
streamlit>=1.24.0
